multi: Fixes INV processing logic
INVs where being sent out to the event queue straight-away with their time set
to the next tricking for the given peer. This implied that, if the node where to
receive a given transaction from that peer, the announcement would still go out as
scheduled.
This has now been fixed by introducing a new event. INVs are scheduled and the simulator
is asked to wake up the node when they are supposed to go out. At that time, the node will
check who may already know about the given transaction and broadcast accordingly