sim: Add an end of simulation for Erlay sims
In contrast to fanout sims, Erlay sims have an recurring event: the reqrecon.
These are scheduled on a rolling timer and, if not dealt with, will make the sim
run forever.

We consider all reconciliation done, for a given node, if it knows all transactions
to be broadcast and all its peers do too.