sim: Properly defines first seen time
Currently, the propagation time is defined as the time between a transaction being created and
it being received by a certain percentage of the network. However, the time the node originator
initially withholds that transaction should not be taken into account. The time should be counted
from when the node broadcasts the transaction.

Account for this as the time the first node receives the transaction minus the link latency