lib: Use Hash{map, set} from hashbrown
The std::collections Hash{map, set} are not the most performance-friendly.
They use SipHash as their hashing algorithm for DoS resistance purposes, but
it makes them slower. hashbrwon has more performant versions of both map and set,
with the compromise of DoS resistance, but that should not be an issue for our
simulator

**Notice this will change any Erlay results having obtained on previous version of the simulator**
This is because changing the hashing function means that the peers being selected for fanout also change
(see `get_fanout_peers` to understand why). The new results should still be accurate, but using a fixed seed
will yield slightly different numbers.