lib: Makes the simulator simulate a single transaction
The simulator was initially designed to allow simulating sending multiple transactions at the same time.
This, however, means having to pass dynamically allocated vectors around and store them in the priority_queue,
which greatly affects the performance of the simulation. We were, however, simulating sending a single transaction,
meaning that a full re-design to get rid of all vector instances could be perform to get better performance while
still being able to use the simulator for our purposes.

This PR gets rid of all vectors in NetworkMessage, plus of the concept of TxId. Given we are sending a single transaction,
a boolean can be used instead. This therefore also changes the way the Erlay messages work. Instead of performing differences
between sets (of 1 or 0 transactions) now we perform differences between booleans, which is equivalent.