lib: pick the source node id before building the network
The node id where the transaction will originate from has always been picked
after the network was built. For the most part, this hasn't mattered much,
given simulations are usually run using the same size network and the same
amount of outbound connections per node. However, if, for instance, the number
of connections per node is changed, that would affect the source node in the
simulation. This is not a big deal, but I think it is best to pick the same
one in cases where we want to compare the simulation outcomes