lib: Makes simulations 100% reproducible
This fixes the last bits of non-reproducibility in the simulator. The only missing
part was related to how peers were iterated when random ones needed to be picked.
Creating iterating over a hashmap would build an iterator whose order depends
on the hashmap random state. However, we were not in control con how that random state
was seeded. Replacing HashMaps with BTreeMaps yields iterators that preserve the order, which
gets rid of the unexpected randomness.

Gets rid of Hashbrown since we are not using HashMaps that heavily anymore.