multi: Makes each peer take care of its own reconciliation loop
Refactor the way the reconciliation round robin is implemented so peers can take care
of stopping their own reconciliation loop once the transaction has been already seen
over than link.

Previously, a single loop was selecting peers in order to do reconciliation, and the simulator
was intercepting schedule messages and dropping them once the node and all its peers knew about
the transaction. This was breaking breaking the simulator layers, which was not too nice.

This will make will also make it easier to test alternative approaches that rely on scheduling
reconciliation for a chosen peer after a given signal