lib: Replies to reconciliation requests on trickle
Instead of replying to reconciliation request straight-away, do so in trickle.
This means that responses will always contain whatever was in the delayed_set.

This may help reduce propagation times, since now a sketches are certain to contain
what the requested node knows, even if they are served less often