multi: Stores all peers in a single collection, and redefines peers
Instead of having two collections, one for outbound peers and one for inbounds,
we store all peers in the same collection and filter as needed. This simplifies
a lot of the handling of both collections at the cost of filtering.

Also, move the is_inbound flag inside the Peer structure. That way we do not need to do a lookup
once we are already holding the Peer structure to check for it.